Fix get definition in versiones processes
